+++ service/msg_service/message.go
@@ -15,27 +15,30 @@ const (
 	PassedMsg = "{{.Name}}你好，你于{{.Date}}提交的{{.Cert}}电子证书领取申请已审核通过，可在【电子证书】中选择证书查看下载"
 	RejectMsg = "{{.Name}}你好，很抱歉你于{{.Date}}提交的{{.Cert}}电子证书领取申请未审核通过，可在【电子证书】中重新申请"
 )
+
 var Msg = map[int]bool{
 	models.Passed: true,
 	models.Reject: true,
 }
 
 type WxMessage struct {
-	Name string
-	Date string
-	Cert string
-	Amount int
+	Name   string
+	Date   string
+	Cert   string
+	Amount float64
 	Openid string
 }
 
 func Send(act int, data *models.C_Apply) {
-	if !Msg[act] { return }
+	if !Msg[act] {
+		return
+	}
 
 	t := time.Unix(data.ApplyDate, 0)
 	m := WxMessage{
-		Name: data.Name,
-		Cert: data.CertName,
-		Date: fmt.Sprintf("%d年%d月%d日", t.Year(), t.Month(), t.Day()),
+		Name:   data.Name,
+		Cert:   data.CertName,
+		Date:   fmt.Sprintf("%d年%d月%d日", t.Year(), t.Month(), t.Day()),
 		Amount: data.PayAmount,
 		Openid: data.WechatID,
 	}
